+++ src/fake-schema/field/directives/fake/index.ts
@@ -1,0 +1,1 @@
+// TODO: move fakeValue here and turn into a class
